+++ src/test/calc_test/parser.ts
@@ -149,7 +149,7 @@ export class Parser {
                 let $scope$sm: Nullable<FAC>;
                 let $$res: Nullable<SUM_$0> = null;
                 if (true
-                    && ($scope$op = this.regexAccept(String.raw`(?:\+|-)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:\+|-)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$sm = this.matchFAC($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.SUM_$0, op: $scope$op, sm: $scope$sm};
@@ -179,7 +179,7 @@ export class Parser {
                 let $scope$sm: Nullable<ATOM>;
                 let $$res: Nullable<FAC_$0> = null;
                 if (true
-                    && ($scope$op = this.regexAccept(String.raw`(?:\*|/)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:\*|/)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$sm = this.matchATOM($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.FAC_$0, op: $scope$op, sm: $scope$sm};
@@ -215,9 +215,9 @@ export class Parser {
                 let $$res: Nullable<ATOM_2> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$val = this.matchSUM($$dpth + 1, $$cr)) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new ATOM_2($scope$val);
@@ -231,7 +231,7 @@ export class Parser {
                 let $scope$val: Nullable<string>;
                 let $$res: Nullable<INT> = null;
                 if (true
-                    && ($scope$val = this.regexAccept(String.raw`(?:[0-9]+)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$val = this.regexAccept(String.raw`(?:[0-9]+)`, "", $$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new INT($scope$val);
                 }
@@ -239,7 +239,7 @@ export class Parser {
             });
     }
     public match_($$dpth: number, $$cr?: ErrorTracker): Nullable<_> {
-        return this.regexAccept(String.raw`(?:\s*)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:\s*)`, "", $$dpth + 1, $$cr);
     }
     public test(): boolean {
         const mrk = this.mark();
@@ -303,10 +303,10 @@ export class Parser {
         }
         return null;
     }
-    private regexAccept(match: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
+    private regexAccept(match: string, mods: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
         return this.run<string>(dpth,
             () => {
-                const reg = new RegExp(match, "y");
+                const reg = new RegExp(match, "y" + mods);
                 const mrk = this.mark();
                 reg.lastIndex = mrk.overallPos;
                 const res = this.tryConsume(reg);