+++ src/test/case_insensitive_test/test.test.ts
@@ -1,0 +1,13 @@
+import { parse } from "./parser";
+
+test(`Case Test 1`, () => {
+    const res = parse("SELECT from");
+    expect(res.errs).toEqual([]);
+    expect(res.ast).not.toBeNull();
+});
+
+test(`Case Test 2`, () => {
+    const res = parse("SeLeCT fRoM");
+    expect(res.errs).toEqual([]);
+    expect(res.ast).not.toBeNull();
+});