+++ src/test/id_test/parser.ts
@@ -67,16 +67,16 @@ export class Parser {
     }
     protected $scope$rule$memo: Map<number, [Nullable<rule>, PosInfo]> = new Map();
     public matchlowercase($$dpth: number, $$cr?: ErrorTracker): Nullable<lowercase> {
-        return this.regexAccept(String.raw`(?:a)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:a)`, "", $$dpth + 1, $$cr);
     }
     public matchUPPERCASE($$dpth: number, $$cr?: ErrorTracker): Nullable<UPPERCASE> {
-        return this.regexAccept(String.raw`(?:b)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:b)`, "", $$dpth + 1, $$cr);
     }
     public match_start_hypen_($$dpth: number, $$cr?: ErrorTracker): Nullable<_start_hypen_> {
-        return this.regexAccept(String.raw`(?:c)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:c)`, "", $$dpth + 1, $$cr);
     }
     public matchnumbers1ab234($$dpth: number, $$cr?: ErrorTracker): Nullable<numbers1ab234> {
-        return this.regexAccept(String.raw`(?:d)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:d)`, "", $$dpth + 1, $$cr);
     }
     public matchrule($$dpth: number, $$cr?: ErrorTracker): Nullable<rule> {
         const fn = () => {
@@ -123,7 +123,7 @@ export class Parser {
                 let $scope$res: Nullable<string>;
                 let $$res: Nullable<rule2> = null;
                 if (true
-                    && ($scope$res = this.regexAccept(String.raw`(?:a)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$res = this.regexAccept(String.raw`(?:a)`, "", $$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.rule2, res: $scope$res};
                 }
@@ -136,7 +136,7 @@ export class Parser {
                 let $scope$cr: Nullable<string>;
                 let $$res: Nullable<rule3> = null;
                 if (true
-                    && ($scope$cr = this.regexAccept(String.raw`(?:b)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$cr = this.regexAccept(String.raw`(?:b)`, "", $$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.rule3, cr: $scope$cr};
                 }
@@ -205,10 +205,10 @@ export class Parser {
         }
         return null;
     }
-    private regexAccept(match: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
+    private regexAccept(match: string, mods: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
         return this.run<string>(dpth,
             () => {
-                const reg = new RegExp(match, "y");
+                const reg = new RegExp(match, "y" + mods);
                 const mrk = this.mark();
                 reg.lastIndex = mrk.overallPos;
                 const res = this.tryConsume(reg);