+++ src/test/memo_time_test/parser.ts
@@ -207,7 +207,7 @@ export class Parser {
                 let $$res: Nullable<E0_1> = null;
                 if (true
                     && ($scope$a = this.matchE1($$dpth + 1, $$cr)) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:a|b)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:a|b)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$b = this.matchE0($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new E0_1($scope$a, $scope$op, $scope$b);
@@ -238,7 +238,7 @@ export class Parser {
                 let $$res: Nullable<E1_1> = null;
                 if (true
                     && ($scope$a = this.matchE2($$dpth + 1, $$cr)) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:c|d)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:c|d)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$b = this.matchE1($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new E1_1($scope$a, $scope$op, $scope$b);
@@ -269,7 +269,7 @@ export class Parser {
                 let $$res: Nullable<E2_1> = null;
                 if (true
                     && ($scope$a = this.matchE3($$dpth + 1, $$cr)) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:e|f)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:e|f)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$b = this.matchE2($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new E2_1($scope$a, $scope$op, $scope$b);
@@ -300,7 +300,7 @@ export class Parser {
                 let $$res: Nullable<E3_1> = null;
                 if (true
                     && ($scope$a = this.matchE4($$dpth + 1, $$cr)) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:g|h)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:g|h)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$b = this.matchE3($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new E3_1($scope$a, $scope$op, $scope$b);
@@ -331,7 +331,7 @@ export class Parser {
                 let $$res: Nullable<E4_1> = null;
                 if (true
                     && ($scope$a = this.matchATOM($$dpth + 1, $$cr)) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:i|j)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:i|j)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$b = this.matchE4($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new E4_1($scope$a, $scope$op, $scope$b);
@@ -372,9 +372,9 @@ export class Parser {
                 let $scope$val: Nullable<E0>;
                 let $$res: Nullable<ATOM_2> = null;
                 if (true
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$val = this.matchE0($$dpth + 1, $$cr)) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new ATOM_2($scope$val);
                 }
@@ -389,7 +389,7 @@ export class Parser {
                         let $scope$val: Nullable<string>;
                         let $$res: Nullable<INT> = null;
                         if (true
-                            && ($scope$val = this.regexAccept(String.raw`(?:[0-9]+)`, $$dpth + 1, $$cr)) !== null
+                            && ($scope$val = this.regexAccept(String.raw`(?:[0-9]+)`, "", $$dpth + 1, $$cr)) !== null
                         ) {
                             $$res = new INT($scope$val);
                         }
@@ -461,10 +461,10 @@ export class Parser {
         }
         return null;
     }
-    private regexAccept(match: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
+    private regexAccept(match: string, mods: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
         return this.run<string>(dpth,
             () => {
-                const reg = new RegExp(match, "y");
+                const reg = new RegExp(match, "y" + mods);
                 const mrk = this.mark();
                 reg.lastIndex = mrk.overallPos;
                 const res = this.tryConsume(reg);