+++ src/test/regex_flags_test/parser.ts
@@ -29,7 +29,7 @@ export class Parser {
     public clearMemos(): void {
     }
     public matchtest($$dpth: number, $$cr?: ErrorTracker): Nullable<test> {
-        return this.regexAccept(String.raw`(?:\p{Alpha}+)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:\p{Alpha}+)`, "", $$dpth + 1, $$cr);
     }
     public test(): boolean {
         const mrk = this.mark();
@@ -93,10 +93,10 @@ export class Parser {
         }
         return null;
     }
-    private regexAccept(match: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
+    private regexAccept(match: string, mods: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
         return this.run<string>(dpth,
             () => {
-                const reg = new RegExp(match, "yu");
+                const reg = new RegExp(match, "yu" + mods);
                 const mrk = this.mark();
                 reg.lastIndex = mrk.overallPos;
                 const res = this.tryConsume(reg);