+++ src/test/setanta_test/parser.ts
@@ -39,7 +39,7 @@
 *                .accept = Acceptor { return <T>(v: ASTVisitor<T>) => v.visitBlockStmt(this); }
 * NuairStmt   := _ 'nuair-a' expr=Expr &gap stmt=NonAsgnStmt
 *                .accept = Acceptor { return <T>(v: ASTVisitor<T>) => v.visitNuairStmt(this); }
-* LeStmt      := _ 'le' &gap id=ID _ 'idir' _ '\('strt=Expr _ ',' end=Expr step={_ ',' step=Expr}? _ '\)' stmt=NonAsgnStmt
+* LeStmt      := _ 'le' &gap id=ID _ 'idir' _ '\(' strt=Expr _ ',' end=Expr step={_ ',' step=Expr}? _ '\)' stmt=NonAsgnStmt
 *                .accept = Acceptor { return <T>(v: ASTVisitor<T>) => v.visitLeStmt(this); }
 * DefnStmt    := _ idstart=@ id=ID idend=@ _ ':=' _ expr=Expr
 *                .accept = Acceptor { return <T>(v: ASTVisitor<T>) => v.visitDefnStmt(this); }
@@ -1079,7 +1079,7 @@ export class Parser {
                 let $$res: Nullable<IfStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:m[áa])`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:m[áa])`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$expr = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
@@ -1098,7 +1098,7 @@ export class Parser {
                 let $$res: Nullable<IfStmt_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:n[oó])`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:n[oó])`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$stmt = this.matchNonAsgnStmt($$dpth + 1, $$cr)) !== null
                 ) {
@@ -1114,10 +1114,10 @@ export class Parser {
                 let $$res: Nullable<BlockStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:{)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:{)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$blk = this.loop<AsgnStmt>(() => this.matchAsgnStmt($$dpth + 1, $$cr), true)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:})`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:})`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new BlockStmt($scope$blk);
                 }
@@ -1132,7 +1132,7 @@ export class Parser {
                 let $$res: Nullable<NuairStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:nuair-a)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:nuair-a)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$expr = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$stmt = this.matchNonAsgnStmt($$dpth + 1, $$cr)) !== null
@@ -1153,20 +1153,20 @@ export class Parser {
                 let $$res: Nullable<LeStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:le)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:le)`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:idir)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:idir)`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$strt = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:,)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:,)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$end = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && (($scope$step = this.matchLeStmt_$0($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$stmt = this.matchNonAsgnStmt($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new LeStmt($scope$id, $scope$strt, $scope$end, $scope$step, $scope$stmt);
@@ -1181,7 +1181,7 @@ export class Parser {
                 let $$res: Nullable<LeStmt_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:,)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:,)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$step = this.matchExpr($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.LeStmt_$0, step: $scope$step};
@@ -1203,7 +1203,7 @@ export class Parser {
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
                     && ($scope$idend = this.mark()) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?::=)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?::=)`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
                     && ($scope$expr = this.matchExpr($$dpth + 1, $$cr)) !== null
                 ) {
@@ -1245,19 +1245,19 @@ export class Parser {
                 let $$res: Nullable<GniomhStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:gn[íi]omh)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:gn[íi]omh)`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && (($scope$args = this.matchCSIDs($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:{)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:{)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$stmts = this.loop<AsgnStmt>(() => this.matchAsgnStmt($$dpth + 1, $$cr), true)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:})`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:})`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new GniomhStmt($scope$id, $scope$args, $scope$stmts);
                 }
@@ -1273,15 +1273,15 @@ export class Parser {
                 let $$res: Nullable<CtlchStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:creatlach)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:creatlach)`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
                     && (($scope$tuis = this.matchCtlchStmt_$0($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:{)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:{)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$gniomhs = this.loop<GniomhStmt>(() => this.matchGniomhStmt($$dpth + 1, $$cr), true)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:})`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:})`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new CtlchStmt($scope$id, $scope$tuis, $scope$gniomhs);
                 }
@@ -1297,7 +1297,7 @@ export class Parser {
                 let $$res: Nullable<CtlchStmt_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:ó)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:ó)`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && ($scope$parentstart = this.mark()) !== null
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
@@ -1314,7 +1314,7 @@ export class Parser {
                 let $$res: Nullable<BrisStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:bris)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:bris)`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = {kind: ASTKinds.BrisStmt, };
                 }
@@ -1327,7 +1327,7 @@ export class Parser {
                 let $$res: Nullable<CCStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:chun-cinn)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:chun-cinn)`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = {kind: ASTKinds.CCStmt, };
                 }
@@ -1341,7 +1341,7 @@ export class Parser {
                 let $$res: Nullable<ToradhStmt> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:toradh)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:toradh)`, "", $$dpth + 1, $$cr) !== null
                     && this.noConsume<gap>(() => this.matchgap($$dpth + 1, $$cr)) !== null
                     && (($scope$exp = this.matchExpr($$dpth + 1, $$cr)) || true)
                 ) {
@@ -1379,7 +1379,7 @@ export class Parser {
                 let $$res: Nullable<And_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\&)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\&)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$trm = this.matchOr($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.And_$0, trm: $scope$trm};
@@ -1413,7 +1413,7 @@ export class Parser {
                 let $$res: Nullable<Or_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\|)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\|)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$trm = this.matchEq($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.Or_$0, trm: $scope$trm};
@@ -1448,7 +1448,7 @@ export class Parser {
                 let $$res: Nullable<Eq_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && ($scope$op = this.regexAccept(String.raw`(?:[!=]=)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$op = this.regexAccept(String.raw`(?:[!=]=)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$trm = this.matchComp($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.Eq_$0, op: $scope$op, trm: $scope$trm};
@@ -1572,7 +1572,7 @@ export class Parser {
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
                     && ($scope$start = this.mark()) !== null
-                    && (($scope$op = this.regexAccept(String.raw`(?:-|!)`, $$dpth + 1, $$cr)) || true)
+                    && (($scope$op = this.regexAccept(String.raw`(?:-|!)`, "", $$dpth + 1, $$cr)) || true)
                     && ($scope$pf = this.matchPostfix($$dpth + 1, $$cr)) !== null
                     && ($scope$end = this.mark()) !== null
                 ) {
@@ -1626,7 +1626,7 @@ export class Parser {
                 let $$res: Nullable<ObjLookups_$0> = null;
                 if (true
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
-                    && this.regexAccept(String.raw`(?:@)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:@)`, "", $$dpth + 1, $$cr) !== null
                     && this.negate(() => this.matchwspace($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.ObjLookups_$0, id: $scope$id};
@@ -1646,10 +1646,10 @@ export class Parser {
                 let $scope$args: Nullable<Nullable<CSArgs>>;
                 let $$res: Nullable<PostOp_1> = null;
                 if (true
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && (($scope$args = this.matchCSArgs($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = {kind: ASTKinds.PostOp_1, args: $scope$args};
                 }
@@ -1662,10 +1662,10 @@ export class Parser {
                 let $scope$expr: Nullable<Expr>;
                 let $$res: Nullable<PostOp_2> = null;
                 if (true
-                    && this.regexAccept(String.raw`(?:\[)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\[)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$expr = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\])`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\])`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = {kind: ASTKinds.PostOp_2, expr: $scope$expr};
                 }
@@ -1691,10 +1691,10 @@ export class Parser {
                 let $$res: Nullable<Atom_1> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$trm = this.matchExpr($$dpth + 1, $$cr)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new Atom_1($scope$trm);
                 }
@@ -1730,17 +1730,17 @@ export class Parser {
                 let $$res: Nullable<GniomhExpr> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:gn[íi]omh)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:gn[íi]omh)`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\()`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\()`, "", $$dpth + 1, $$cr) !== null
                     && (($scope$args = this.matchCSIDs($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\))`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\))`, "", $$dpth + 1, $$cr) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:{)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:{)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$stmts = this.loop<AsgnStmt>(() => this.matchAsgnStmt($$dpth + 1, $$cr), true)) !== null
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:})`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:})`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new GniomhExpr($scope$args, $scope$stmts);
                 }
@@ -1754,10 +1754,10 @@ export class Parser {
                 let $$res: Nullable<ListLit> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\[)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\[)`, "", $$dpth + 1, $$cr) !== null
                     && (($scope$els = this.matchCSArgs($$dpth + 1, $$cr)) || true)
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:\])`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\])`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new ListLit($scope$els);
                 }
@@ -1790,7 +1790,7 @@ export class Parser {
                 let $$res: Nullable<CSArgs_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:,)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:,)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$exp = this.matchExpr($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.CSArgs_$0, exp: $scope$exp};
@@ -1820,7 +1820,7 @@ export class Parser {
                 let $$res: Nullable<CSIDs_$0> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:,)`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:,)`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$id = this.matchID($$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = {kind: ASTKinds.CSIDs_$0, id: $scope$id};
@@ -1839,7 +1839,7 @@ export class Parser {
                     && this.match_($$dpth + 1, $$cr) !== null
                     && this.negate(() => this.matchID_$0($$dpth + 1, $$cr)) !== null
                     && ($scope$start = this.mark()) !== null
-                    && ($scope$id = this.regexAccept(String.raw`(?:[a-zA-Z_áéíóúÁÉÍÓÚ][a-zA-Z_áéíóúÁÉÍÓÚ0-9]*)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$id = this.regexAccept(String.raw`(?:[a-zA-Z_áéíóúÁÉÍÓÚ][a-zA-Z_áéíóúÁÉÍÓÚ0-9]*)`, "", $$dpth + 1, $$cr)) !== null
                     && ($scope$end = this.mark()) !== null
                 ) {
                     $$res = new ID($scope$start, $scope$id, $scope$end);
@@ -1867,7 +1867,7 @@ export class Parser {
                 let $$res: Nullable<Bool> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && ($scope$bool = this.regexAccept(String.raw`(?:f[ií]or|br[eé]ag)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$bool = this.regexAccept(String.raw`(?:f[ií]or|br[eé]ag)`, "", $$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new Bool($scope$bool);
                 }
@@ -1880,7 +1880,7 @@ export class Parser {
                 let $$res: Nullable<Neamhni> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && this.regexAccept(String.raw`(?:neamhn[ií])`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:neamhn[ií])`, "", $$dpth + 1, $$cr) !== null
                 ) {
                     $$res = new Neamhni();
                 }
@@ -1894,7 +1894,7 @@ export class Parser {
                 let $$res: Nullable<Int> = null;
                 if (true
                     && this.match_($$dpth + 1, $$cr) !== null
-                    && ($scope$int = this.regexAccept(String.raw`(?:-?[0-9]+(?:\.[0-9]+)?)`, $$dpth + 1, $$cr)) !== null
+                    && ($scope$int = this.regexAccept(String.raw`(?:-?[0-9]+(?:\.[0-9]+)?)`, "", $$dpth + 1, $$cr)) !== null
                 ) {
                     $$res = new Int($scope$int);
                 }
@@ -1930,9 +1930,9 @@ export class Parser {
                 let $$res: Nullable<Teacs_$0_1> = null;
                 if (true
                     && ($scope$start = this.mark()) !== null
-                    && this.regexAccept(String.raw`(?:\')`, $$dpth + 1, $$cr) !== null
-                    && ($scope$val = this.regexAccept(String.raw`(?:([^\'\\]|\\.)*)`, $$dpth + 1, $$cr)) !== null
-                    && this.regexAccept(String.raw`(?:\')`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:\')`, "", $$dpth + 1, $$cr) !== null
+                    && ($scope$val = this.regexAccept(String.raw`(?:([^\'\\]|\\.)*)`, "", $$dpth + 1, $$cr)) !== null
+                    && this.regexAccept(String.raw`(?:\')`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$end = this.mark()) !== null
                 ) {
                     $$res = {kind: ASTKinds.Teacs_$0_1, start: $scope$start, val: $scope$val, end: $scope$end};
@@ -1949,9 +1949,9 @@ export class Parser {
                 let $$res: Nullable<Teacs_$0_2> = null;
                 if (true
                     && ($scope$start = this.mark()) !== null
-                    && this.regexAccept(String.raw`(?:")`, $$dpth + 1, $$cr) !== null
-                    && ($scope$val = this.regexAccept(String.raw`(?:([^"\\]|\\.)*)`, $$dpth + 1, $$cr)) !== null
-                    && this.regexAccept(String.raw`(?:")`, $$dpth + 1, $$cr) !== null
+                    && this.regexAccept(String.raw`(?:")`, "", $$dpth + 1, $$cr) !== null
+                    && ($scope$val = this.regexAccept(String.raw`(?:([^"\\]|\\.)*)`, "", $$dpth + 1, $$cr)) !== null
+                    && this.regexAccept(String.raw`(?:")`, "", $$dpth + 1, $$cr) !== null
                     && ($scope$end = this.mark()) !== null
                 ) {
                     $$res = {kind: ASTKinds.Teacs_$0_2, start: $scope$start, val: $scope$val, end: $scope$end};
@@ -1963,7 +1963,7 @@ export class Parser {
         return this.loop<wspace>(() => this.matchwspace($$dpth + 1, $$cr), true);
     }
     public matchwspace($$dpth: number, $$cr?: ErrorTracker): Nullable<wspace> {
-        return this.regexAccept(String.raw`(?:(?:\s|>--(?:(?!--<).)*(--<|\n|$)))`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:(?:\s|>--(?:(?!--<).)*(--<|\n|$)))`, "", $$dpth + 1, $$cr);
     }
     public matchgap($$dpth: number, $$cr?: ErrorTracker): Nullable<gap> {
         return this.choice<gap>([
@@ -1976,22 +1976,22 @@ export class Parser {
         return this.matchwspace($$dpth + 1, $$cr);
     }
     public matchgap_2($$dpth: number, $$cr?: ErrorTracker): Nullable<gap_2> {
-        return this.regexAccept(String.raw`(?:[^a-zA-Z0-9áéíóúÁÉÍÓÚ])`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:[^a-zA-Z0-9áéíóúÁÉÍÓÚ])`, "", $$dpth + 1, $$cr);
     }
     public matchgap_3($$dpth: number, $$cr?: ErrorTracker): Nullable<gap_3> {
-        return this.regexAccept(String.raw`(?:$)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:$)`, "", $$dpth + 1, $$cr);
     }
     public matchPlusMinus($$dpth: number, $$cr?: ErrorTracker): Nullable<PlusMinus> {
-        return this.regexAccept(String.raw`(?:\+|-)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:\+|-)`, "", $$dpth + 1, $$cr);
     }
     public matchAsgnOp($$dpth: number, $$cr?: ErrorTracker): Nullable<AsgnOp> {
-        return this.regexAccept(String.raw`(?:=|\+=|\*=|-=|%=|\/=|\/\/=)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:=|\+=|\*=|-=|%=|\/=|\/\/=)`, "", $$dpth + 1, $$cr);
     }
     public matchMulDiv($$dpth: number, $$cr?: ErrorTracker): Nullable<MulDiv> {
-        return this.regexAccept(String.raw`(?:\*|\/\/|%|\/)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:\*|\/\/|%|\/)`, "", $$dpth + 1, $$cr);
     }
     public matchCompare($$dpth: number, $$cr?: ErrorTracker): Nullable<Compare> {
-        return this.regexAccept(String.raw`(?:<=|>=|<|>)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:<=|>=|<|>)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword> {
         return this.choice<Keyword>([
@@ -2007,31 +2007,31 @@ export class Parser {
         ]);
     }
     public matchKeyword_1($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_1> {
-        return this.regexAccept(String.raw`(?:m[áa])`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:m[áa])`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_2($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_2> {
-        return this.regexAccept(String.raw`(?:n[oó])`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:n[oó])`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_3($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_3> {
-        return this.regexAccept(String.raw`(?:nuair-a)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:nuair-a)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_4($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_4> {
-        return this.regexAccept(String.raw`(?:f[ií]or|br[eé]ag)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:f[ií]or|br[eé]ag)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_5($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_5> {
-        return this.regexAccept(String.raw`(?:gn[ií]omh)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:gn[ií]omh)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_6($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_6> {
-        return this.regexAccept(String.raw`(?:chun-cinn)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:chun-cinn)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_7($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_7> {
-        return this.regexAccept(String.raw`(?:neamhn[ií])`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:neamhn[ií])`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_8($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_8> {
-        return this.regexAccept(String.raw`(?:toradh)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:toradh)`, "", $$dpth + 1, $$cr);
     }
     public matchKeyword_9($$dpth: number, $$cr?: ErrorTracker): Nullable<Keyword_9> {
-        return this.regexAccept(String.raw`(?:creatlach)`, $$dpth + 1, $$cr);
+        return this.regexAccept(String.raw`(?:creatlach)`, "", $$dpth + 1, $$cr);
     }
     public test(): boolean {
         const mrk = this.mark();
@@ -2095,10 +2095,10 @@ export class Parser {
         }
         return null;
     }
-    private regexAccept(match: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
+    private regexAccept(match: string, mods: string, dpth: number, cr?: ErrorTracker): Nullable<string> {
         return this.run<string>(dpth,
             () => {
-                const reg = new RegExp(match, "y");
+                const reg = new RegExp(match, "y" + mods);
                 const mrk = this.mark();
                 reg.lastIndex = mrk.overallPos;
                 const res = this.tryConsume(reg);