+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `+` operator return type now specifies non-empty arrays (#49).
 - Grammars now are checked for any Typescript keywords before parser generation, to avoid
   generating invalid TypeScript (#44).
+- Added support for regex modifiers including case insensitivity and multiline mode (#48).
 
 ### Fixed